approach1:introduce defaults
