+++ src/datasources/results/ResultsDataSource.ts
@@ -1,8 +1,9 @@
 import { DataFrameDTO, DataQueryRequest, DataSourceInstanceSettings, TestDataSourceResponse } from '@grafana/data';
 import { BackendSrv, TemplateSrv, getBackendSrv, getTemplateSrv } from '@grafana/runtime';
 import { DataSourceBase } from 'core/DataSourceBase';
-import { ResultsQuery } from './types';
+import { ResultsQuery } from './types/types';
 import { QueryResultsDataSource } from './data-sources/query-results/QueryResultsDataSource';
+import { QueryResults } from './types/QueryResults.types';
 
 export class ResultsDataSource extends DataSourceBase<ResultsQuery> {
   public defaultQuery: Partial<ResultsQuery> & Omit<ResultsQuery, 'refId'>;
@@ -16,13 +17,13 @@ export class ResultsDataSource extends DataSourceBase<ResultsQuery> {
   ) {
     super(instanceSettings, backendSrv, templateSrv);
     this.queryResultsDataSource = new QueryResultsDataSource(instanceSettings, backendSrv, templateSrv);
-    this.defaultQuery = this.queryResultsDataSource.defaultQuery;
+    this.defaultQuery = this.queryResultsDataSource.defaultQuery as QueryResults;
   }
   
   baseUrl = this.instanceSettings.url + '/nitestmonitor';
 
   async runQuery(query: ResultsQuery, options: DataQueryRequest): Promise<DataFrameDTO> {
-    return this.queryResultsDataSource.runQuery(query as ResultsQuery, options);
+    return this.queryResultsDataSource.runQuery(query as QueryResults, options);
   }
 
   shouldRunQuery(query: ResultsQuery): boolean {
