+++ src/datasources/results/components/ResultsQueryEditor.tsx
@@ -1,8 +1,11 @@
-import React, { useCallback } from 'react';
-import { QueryEditorProps } from '@grafana/data';
+import React, { useCallback, useState } from 'react';
+import { QueryEditorProps, SelectableValue } from '@grafana/data';
 import { ResultsDataSource } from '../ResultsDataSource';
-import { ResultsQuery } from '../types';
+import { QueryType, ResultsQuery } from '../types/types';
 import { QueryResultsEditor } from './editors/query-results/QueryResultsEditor';
+import { QueryResults } from '../types/QueryResults.types';
+import { InlineField, RadioButtonGroup, VerticalGroup } from '@grafana/ui';
+import { ResultsDefaults, StepsDefaults } from '../defaults';
 
 type Props = QueryEditorProps<ResultsDataSource, ResultsQuery>;
 
@@ -18,10 +21,30 @@ export function ResultsQueryEditor({ query, onChange, onRunQuery, datasource }: 
     },[onChange, onRunQuery]
   );
 
+  const handleQueryTypeChange = useCallback((value: QueryType): void => {
+    if (value === QueryType.Results) {
+      handleQueryChange({ ...query, queryType: QueryType.Results, ...ResultsDefaults}, true);
+    }
+    if (value === QueryType.Steps) {
+      handleQueryChange({ ...query, queryType: QueryType.Steps, ...StepsDefaults }, true);
+    }
+  }, [query, handleQueryChange]);
+
   return (
-    <QueryResultsEditor
-      query={query} 
-      handleQueryChange={handleQueryChange}
-    />
+    <VerticalGroup>
+      <InlineField label="Query Type" labelWidth={18}>
+        <RadioButtonGroup
+          options={Object.values(QueryType).map(value => ({ label: value, value })) as SelectableValue[]}
+          value={query.queryType}
+          onChange={handleQueryTypeChange}
+        />
+      </InlineField>
+      {query.queryType === QueryType.Results && (
+        <QueryResultsEditor 
+          query={query as QueryResults} 
+          handleQueryChange={handleQueryChange} 
+        />
+      )}
+    </VerticalGroup>
   );
 }
